[WIP] Store: add more robust error reporting
In order to avoid dynamic memory allocation,
the store makes use of the single pointer passed in by the caller as a
buffer for both returning data and reporting errors.
This requires the use of a magic canary value to indicate errors.
This value needs to be different from any legitimate succesfull
responses.
Previously false was used in certain cases, which is a potential
legitimate value. This has now been replaced with the canary,
which is set to INT16_MAX.