Store fix finding latest sample
Finding the latest sample didn't actually work, as we were using random
grabage (the first couple bytes of the buffer) as the id instead of
last_id, and kept looping until we found a sample that didn't exist
instead of one that does.
Turns out we keep our state "clean" after deletions, so last_id always
points to the latest Sample - looping is useless.