Refactor the date support
Modify the ds3231 driver to use the standard tm struct for setting and
getting the time.
Add convenience functions for manipulating the UTC time and converting
it to the UNIX epoch in utc_time.h. This includes source adapted from
gmtime and mkgmtime.
Fix sampling-semsors to properly convert tm struct to epoch properly.