Fix buffer-overflow in res_routes
We were writing to a buffer in a while loop, with the guard checking a
least a single byte was remaining.
Unfortunately we were righting more than one byte per iteration, which
could lead to a potential buffer overflow.

To make things worse, the limiter in calls to snprintf aimed at
preventing buffer overflowas was not set correctly, allowing buffer
overflows.