Revert "Variable Paranoia. Move all store vars to bss"
This reverts commit 88bfad2bab00ace81e98b2954b5487e01b40609e.

This commit was made as a last ditch effort to fix the nodes on the last
day of deploymenent.
After further research documentation it has come to
light that all of the variables that were stack allocated could remain
so safely as the stack is only lost when a process yields.
Allocating the variables / buffers on the stack save 190 bytes of ram
(!), which is the main reason for reverting this.