Refactor Store to not run in it's own process
The store previsouly ran in it's own process,
any synchronously handled events.
It was designed this way to avoid and race / thread safety issues,
but Contiki protothreads are entirely cooperative.
They can only yield when they choose to do so,
and only one can run at a given time.
The store can therefore safely not be in it's own process.