Remove microlog requirement
Micrologs were required (on the Z1 platform) to overwrite files.
The only file we ever overwrite is the config, we now delete it before
overwriting it.
This might cause a bit more garbage collection to take
place, but it's a small price to pay for freeing up lots of flash /
complexity.