Store: Add support for files with trailing NULL B
Coffee does not support files with trailing NULL bytes properly,
the null byte is remembered properly only as long as the file is cached.
As soon as the file needs to be found in the flash again,
trailing null bytes will not be read.

The Store now appends a magic canary value to every file,
and then strips it when reading the files back.