Change config filename from sampleconfig to config
The maximum length of files was set to 8,
and sampleconfig was larger than that. Seems it may have caused some
issues with file reservations.

As an unintended side effect, it now lets us build with debugging turned
off! (How broken is the compiler... remove 6 bytes from a string literal
and it changes everything).