CC2538: Fix external flash support
External flash is supported through the COFFEE_CONF_CUSTOM_PORT macro,
which when used disables the program flash based Coffee.
The linker script directly uses COFFEE_START and COFFEE_SIZE to reserve
space for the program flash based Coffee.

Consequently, space is reserved in program flash even for external flash
chips, and things break when the linker script attempts to reserve
several MB of space.

This change adds a new macro, CC2538_FLASH_OFFSET, that is used by the
linker script.
If the program flash based Coffee is used, it is set to COFFEE_SIZE,
else set to 0.