Store: Reduce required filename length
Set the max filename length from the one decalred in Coffee instead of
hardcoding it.

Reduce the config filename to 5 characters (so it fits in 6 with the
null byte).

Remove the "r" prefix from sample files so that they're one character
shorter (5 chars allows us to store 99999 samples - more than plenty).

Add 2 static asserts to check that the filename length set in Coffee is
long enough for both the config, and the maximum number of samples we
ever store.