Coffee: Tweak number of cached files / fd set
Reduce the number of files that can be openned at once, (we only ever
use one at a time).

Increase the number of files cached on the muntjac platform ito 3 as we have
plenty of ram to space (should allow all samples to be read from RAM if
we sample every 20min and fetch every hour).

Reduce the number of of files cached on z1-{feshie,wais} to 1 to reduce
RAM usage.