CC1120: Zoul: Stay in LPM0 when radio is ON
Add a lpm_register_peripheral() callback in the zoul CC1120 arch driver,
to prevent the CC2538 from entering LPM{1,2} when the CC1120 is on.

This resolves a latency issue, where the average latency of a pin was
over 1000ms. It is now back down to 230ms.

We presume the extra latency is caused by increased ISR latency when the
CC2538 has to leave LPM{1,2} to return to LPM0.