ContikiMac: Schedule next cycle after last CCA check
After the last CCA check, ContikiMac schedules another wake up and
yields.
When it wakes up again, it just schedules the next cycle check.

This wastes power and time, so once we are done with the last CCA check,
we immediately schedule the next cycle check and yield.
This saves us from performaning an extra yeild / sleep / wakeup cycle.

On a power meter, the 3rd spike seen after the first 2 CCA checks
dissapears. On a muntjac, this reduces our average power consumption by
~0.4mA.